ai/requirements.txt
torch==2.0.1
transformers==4.30.2
pytorch-lightning==2.0.6
pandas==2.0.3
numpy==1.24.4
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
fastapi==0.100.0
uvicorn==0.23.1
pydantic==2.0.3
python-dotenv==1.0.0
pyyaml==6.0.1
mlflow==2.5.0
onnx==1.14.0
onnxruntime==1.15.1
jupyter==1.0.0
openai==0.27.8
tqdm==4.65.0 